PodOperator.restart()
