+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/resource/ZookeeperLeaderFinder.java
@@ -114,7 +114,7 @@ public class ZookeeperLeaderFinder {
     protected PemKeyCertOptions keyCertOptions(Secret coCertKeySecret) {
         CertAndKey coCertKey = Ca.asCertAndKey(coCertKeySecret, "cluster-operator.key", "cluster-operator.crt");
         if (coCertKey == null) {
-            throw missingSecretFuture(coCertKeySecret.getMetadata().getNamespace(), coCertKeySecret.getMetadata().getName());
+            throw StatefulSetOperator.missingSecretFuture(coCertKeySecret.getMetadata().getNamespace(), coCertKeySecret.getMetadata().getName());
         }
         CertificateFactory x509 = x509Factory();
         try {
@@ -140,7 +140,7 @@ public class ZookeeperLeaderFinder {
         Future<Secret> clusterCaKeySecretFuture = secretOperator.getAsync(namespace, clusterCaSecretName);
         return clusterCaKeySecretFuture.compose(clusterCaCertificateSecret -> {
             if (clusterCaCertificateSecret  == null) {
-                return Future.failedFuture(missingSecretFuture(namespace, clusterCaSecretName));
+                return Future.failedFuture(StatefulSetOperator.missingSecretFuture(namespace, clusterCaSecretName));
             }
             try {
                 NetClientOptions netClientOptions = clientOptions(coKeySecret, clusterCaCertificateSecret);
@@ -151,11 +151,6 @@ public class ZookeeperLeaderFinder {
         });
 
     }
-
-    static RuntimeException missingSecretFuture(String namespace, String secretName) {
-        return new RuntimeException("Secret " + namespace + "/" + secretName + " does not exist");
-    }
-
     private Future<Integer> zookeeperLeader(String cluster, String namespace, List<Pod> pods,
                                             NetClientOptions netClientOptions) {
         Future<Integer> result = Future.future();
